backend/requirements.txt
fastapi
uvicorn
requests
internetarchive
PyJWT